методы трансляции

+++ 4 семестр/Методы трансляции/testMT/testMT/lex_yy.c
@@ -1025,7 +1025,7 @@ static int input()
 
 		else
 			{ /* need more input */
-			int offset = yy_c_buf_p - yytext_ptr;
+			long offset = yy_c_buf_p - yytext_ptr;
 			++yy_c_buf_p;
 
 			switch ( yy_get_next_buffer() )
